// src/app/(main)/dashboard/Dashboard.interface.ts
export interface Item {
	id: number;
	object_name: string;
	object_type: 3;
	name: string;
	last_name: string;
	address: string;
	id_number: string;
	email: string;
	mobile: string;
	time_from: string;
	time_to: string;
	discount: null;
	facebook: string;
	instagram: string;
	description: string;
	images: string;
}